requirements.txt
django==1.5.1
mysql-python
pytz

South
django-cities-light